Add CommandPoolGuard class to simplify temporary command pool creation

+++ src/client/renderer/RendererInit.cpp
@@ -372,7 +372,7 @@ namespace VE
         VkCommandPoolCreateInfo graphicsCommandPoolCreateInfo = {
             .sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO,
             .flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT,
-            .queueFamilyIndex = static_cast<uint32_t>(graphicsQueueFamilyIndex)};
+            .queueFamilyIndex = graphicsQueueFamilyIndex};
 
         vkCheck(vkCreateCommandPool(device, &graphicsCommandPoolCreateInfo, nullptr, &commandPool), {'V', 208});
     }
